Cleanup closes #3
