+++ TM.Digital.Client/Resources/Templates/WaitScreenResources.xaml
@@ -7,43 +7,41 @@
                     mc:Ignorable="d">
 
     <DataTemplate DataType="{x:Type wait:WaitingGameScreenViewModel}">
-        
-            <Grid Grid.Row="0" >
-                <Grid.RowDefinitions>
-                    <RowDefinition />
-                <RowDefinition Height="Auto"/>
-                <RowDefinition Height="Auto"/>
+
+        <Grid Grid.Row="0">
+            <Grid.RowDefinitions>
+                <RowDefinition />
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
                 <RowDefinition />
             </Grid.RowDefinitions>
-                <Grid.Background>
-                    <SolidColorBrush Color="White" Opacity="0.8" />
-                </Grid.Background>
-                <Border Grid.Row="1" VerticalAlignment="Center" HorizontalAlignment="Center" Padding="60,10" 
+            <Grid.Background>
+                <SolidColorBrush Color="White" Opacity="0.8" />
+            </Grid.Background>
+            <Border Grid.Row="1" VerticalAlignment="Center" HorizontalAlignment="Center" Padding="60,10"
                         Margin="15" BorderBrush="Black" BorderThickness="1" Background="White">
-                    <Border.Effect>
-                        <DropShadowEffect ShadowDepth="3" Opacity="0.8"></DropShadowEffect>
-                    </Border.Effect>
-                    <TextBlock Text="{Binding InitialMessage}" VerticalAlignment="Center" HorizontalAlignment="Center"
+                <Border.Effect>
+                    <DropShadowEffect ShadowDepth="3" Opacity="0.8"></DropShadowEffect>
+                </Border.Effect>
+                <TextBlock Text="{Binding InitialMessage}" VerticalAlignment="Center" HorizontalAlignment="Center"
                                Margin="10"
                                FontSize="25" FontWeight="Bold"></TextBlock>
-                </Border>
-                <Button Grid.Row="2" Command="{Binding StartGameCommand}" Content="Start Game" VerticalAlignment="Top" HorizontalAlignment="Center"
+            </Border>
+            <Button Grid.Row="2" Command="{Binding StartGameCommand}" Content="Start Game" VerticalAlignment="Top" HorizontalAlignment="Center"
                         Visibility="{Binding IsOwner, Converter={StaticResource boolToVisConverter}}"
                         Padding="20"
                         FontSize="20" FontWeight="Bold"></Button>
-                <ListBox Grid.Row="3" Margin="5" ItemsSource="{Binding IncommingMessages}" IsEnabled="False" HorizontalAlignment="Center"
+            <ListBox Grid.Row="3" Margin="5" ItemsSource="{Binding IncommingMessages}" IsEnabled="False" HorizontalAlignment="Center"
                          VerticalAlignment="Top"
                          BorderBrush="Transparent"
-                         Background="Transparent"
-                         >
-                    <ListBox.ItemTemplate>
-                        <DataTemplate>
+                         Background="Transparent">
+                <ListBox.ItemTemplate>
+                    <DataTemplate>
                         <TextBlock Text="{Binding}" Foreground="Black" FontSize="18"
                                    FontStyle="Italic"></TextBlock>
-                        </DataTemplate>
-                    </ListBox.ItemTemplate>
-                </ListBox>
-            </Grid>
-        
+                    </DataTemplate>
+                </ListBox.ItemTemplate>
+            </ListBox>
+        </Grid>
     </DataTemplate>
 </ResourceDictionary>
