+++ TM.Digital.Client/Screens/Main/MainWindow.xaml
@@ -3,7 +3,7 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:TM.Digital.Client" 
+        xmlns:local="clr-namespace:TM.Digital.Client"
         xmlns:main="clr-namespace:TM.Digital.Client.Screens.Main"
         mc:Ignorable="d"
         WindowStartupLocation="CenterScreen"
