+++ TM.Digital.Client/Services/IApiProxy.cs
@@ -8,8 +8,11 @@ namespace TM.Digital.Client.Services
     public interface IApiProxy
     {
         Task<Player> JoinGame(Guid selectedSessionGameSessionId, string playerName);
+
         Task<GameSessions> GetGameSessions();
+
         Task<GameSessionInformation> CreateNewGame(string playerName, int numberOfPlayers);
+
         Task<bool> StartGame(Guid sessionGameSessionId);
     }
 }
